requirements.txt
Fabric3
paho-mqtt